+++ src/pages/login/Login.tsx
@@ -6,7 +6,6 @@ import Header from '../../components/header/Header'
 import jwt_Decode from 'jwt-decode'
 import { onSetCookieHandler, onSetLocalStorageHandler } from '../../util/cookie'
 import {
-  KakaoLoginBtn,
   KakaoLoginImg,
   LoginBtn,
   LoginContainer,
@@ -18,9 +17,6 @@ import {
 } from './loginSt'
 function Login() {
   const KAKAO_AUTH_URL = `https://kauth.kakao.com/oauth/authorize?client_id=${process.env.REACT_APP_KAKAO_REST_API_KEY}&redirect_uri=${process.env.REACT_APP_KAKAO_REDIRECT_URI}&response_type=code`
-
-  const googleLoginUrl = `https://accounts.google.com/o/oauth2/v2/auth?client_id=${process.env.REACT_APP_GOOGLE_CLIENT_ID}&response_type=token&redirect_uri=${process.env.REACT_APP_GOOGLE_REDIRECT_URI}&scope=https://www.googleapis.com/auth/userinfo.profile`
-  const naverLoginUrl = `https://nid.naver.com/oauth2.0/authorize?response_type=code&client_id=${process.env.REACT_APP_NAVER_CLIENT_ID}&redirect_uri=${process.env.REACT_APP_NAVER_REDIRECT_URI}&state=random_string`
 
   const [id, setId] = useState('')
   const [password, setPassword] = useState('')
@@ -41,12 +37,10 @@ function Login() {
         onSetLocalStorageHandler('nickname', nickname)
         onSetLocalStorageHandler('refresh', refreshToken)
         onSetLocalStorageHandler('userInfo', decodeUserInfo)
-        alert(res.data.message)
         navigate('/recommend')
-        alert(nickname + "님 환영합니다.")
       })
       .catch(() => {
-        alert('아이디 비밀번호가 일치하지 않습니다.')
+        alert('아이디 또는 비밀번호가 일치하지 않습니다.')
       })
   }
 
@@ -54,24 +48,11 @@ function Login() {
     window.location.href = KAKAO_AUTH_URL
   }
 
-  const onGoogleLoginHanlder = () => {
-    window.location.assign(googleLoginUrl)
-  }
-  const onNaverLoginHandler = () => {
-    window.location.href = naverLoginUrl
-  }
   return (
     <>
       <Header />
       <LoginContainer>
         <p>Log in</p>
-        {/* <span
-          onClick={() => {
-            navigate('/signup')
-          }}
-        >
-          MOOD 회원이 아니신가요?
-        </span> */}
         <LoginContainerForm>
           <div>
             <LoginLabelDiv>
@@ -114,9 +95,6 @@ function Login() {
           />
         </LoginSocialContainer>
       </LoginContainer>
-
-      {/* <button onClick={onGoogleLoginHanlder}>구글 로그인</button>
-      <button onClick={onNaverLoginHandler}>네이버 로그인</button> */}
     </>
   )
 }
